gen_defines: output the interrupt level of a node
Currently it is tedious to know the level of an interrupt for
a node in C. One would have to go through a very complex and
error prone macros to check if there's a parent interrupt
controller & if the controller has an interrupt number and thus
not a pseudo interrupt controller like the one found in
`rv32m1`. The level of a node is required to encode the
Zephyr's multi-level interrupt number

Since it is easier to do it in the `gen_defines` script, let's
do the heavy lifting there so that we can introduce new DT
macros to get the interrupt level very easily later.

(cherry picked from commit 450a66fa3632f2f0e260cb13578798eddf6d8974)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
GitOrigin-RevId: 450a66fa3632f2f0e260cb13578798eddf6d8974
Change-Id: I56119c32ad22e5e2c40a993267726b5cba1e8c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5233902
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>