modem: chat: Patch invalid dereference
The chat module contains an array of three lists of matches,
one of which are static, two of which are contained within the
currently running script. The current match, which is an
object stored in one of the three lists, is stored in its own
pointer in the chat module context.

A memory error occurs when the script is stopped, while the
chat module is using one of the matches stored withing the
script. This commit clears the match pointer when the script
is stopped if the match is stored within the script.

(cherry picked from commit 3ebe8142945e53828c1a6c91f4b134adece4af33)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki@arge-andreasen.me>
GitOrigin-RevId: 3ebe8142945e53828c1a6c91f4b134adece4af33
Change-Id: Ib89f96a28027b89a15c50edeeb530e5e87ccec8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5168410
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>