dts-bindings: pinctrl: renesas_ra: enabled config of i/o ports 4-7
The RA_PINCFG macro is used to generate a value that can be written
directly to the pin function select register. In addition to the pin
function this value also contains port and pin number information,
located in bit fields that are unused by the register. The bit field
used to store the port information consists of 3-bits. However, a typo in
the mask definition limited the field to two bits meaning only ports 0-3
could be configured. This patch resolves the issue, allowing ports 0-7 to
be configured. If the port is greater than 7 another field (port4) is used
to store an additional bit (allowing an additional 8 ports to be
supported). However, use of this field has not yet been implemented.

(cherry picked from commit b1a15718aa9990c8820c2f426d7a559b88f74e29)

Original-Signed-off-by: Ian Morris <ian.d.morris@outlook.com>
GitOrigin-RevId: b1a15718aa9990c8820c2f426d7a559b88f74e29
Change-Id: If0264aed1b628205879ba2d9b77f3b923ac17bdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5279896
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>