net: tcp: Install a last ack timer in passive close
If we are in a passive close state, then it is possible that
the ack we are waiting is lost or we do not accept the one peer
sent to us because of some earlier out of memory issue.
So install a timer (using by default the FIN timer value) to
close the connection if the last ack is not received on time.

(cherry picked from commit 40215e07a39305125d47f04f6b4a498b71d79c43)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 40215e07a39305125d47f04f6b4a498b71d79c43
Change-Id: Ie555a370c8f38ea85e12ef07f52ecaea8161f057
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5180690
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>