west: runners: blackmagicprobe: flash hex_file
The cfg.elf_file is not signed, so `west flash` with a sysbuild will
not run the application if MCUboot image signature checking is on. Using
the cfg.hex_file for `bmp_flash` resolves since as the signed hex is
used when using sysbuild. Symbols are not required for flashing so the
switching from elf to hex is not an issue.

(cherry picked from commit 10183797a125329f9b93e22b151e00a00d8d984a)

Original-Signed-off-by: John Whittington <git@jbrengineering.co.uk>
GitOrigin-RevId: 10183797a125329f9b93e22b151e00a00d8d984a
Change-Id: I856cd50945568670839220f1e2f6df9c1bf51f5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5242686
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>