drivers: ieee802154: fix nRF5 Rx error handling
The current implementation implicitly assumes that if the device is
configured to have the capability of acting as a CSL endpoint then in
case a delayed reception with matching ID finishes with a timeout no
action is needed. This assumption is correct when RxOnWhenIdle mode is
disabled because the transition to sleep is done automatically by the
driver below. However, it's wrong when RxOnWhenIdle is enabled. This
commit fixes that case by adding a call to event handler that notifies
the higher layer about the event and allows it to transition to RxOff if
needed.

(cherry picked from commit 8f9a86cf5e19d466a2a13a2f34ee35230f17cfc4)

Original-Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>
GitOrigin-RevId: 8f9a86cf5e19d466a2a13a2f34ee35230f17cfc4
Change-Id: I1757213e8b392a845cf8d8ddf7c024970c65be94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5220370
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>