dts/riscv: remove the `timebase-frequency` property
The `timebase-frequency` is not defined by any of the YAML binding files.
There was a discussion in #37420 to add this property, but in the end it
was rejected. This resulted in the #37685 feature request being created.

As of now, this property is not documented anywhere so this commit removes
it from the RISC-V devicetrees, as RISC-V is the only architecture that is
currently defining it - and even in RISC-V not all platforms do that.

(cherry picked from commit 17670be2cca6f146d7d2e9d81b215cb35871ee4c)

Original-Signed-off-by: Filip Kokosinski <fkokosinski@antmicro.com>
GitOrigin-RevId: 17670be2cca6f146d7d2e9d81b215cb35871ee4c
Change-Id: I1b3e4e9bfd743a324aefbbc51171d1ad21489c5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5279950
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>