Bluetooth: Mesh: Call bt_mesh_send_cb.end cb by the end of adv
Before this change, the bt_mesh_send_cb.end callback was called when all
references to the adv were removed. If an implementation kept more
references to the adv buffer after calling `bt_mesh_adv_send`, the end
callback would not be called when the advertiser finished advertising
this adv.

With this change, the end callback is always called by the advertiser
when the advertisement is finished regardless of the number of
references. This allows an implementation to keep the adv buffer for the
future use. As an example, pb_adv.c keeps advs for retransmission.

(cherry picked from commit 9af051e3499fccdb1898748d2afd8dcd06d013f2)

Original-Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>
GitOrigin-RevId: 9af051e3499fccdb1898748d2afd8dcd06d013f2
Change-Id: I659e3978d45ac168f0f957c7e19b62a971b809d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5244003
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>