drivers: pinctrl: renesas: fix ipsr generation for S4
Avoid unexpected memory access in cases where the IPSR has an odd
register index. In the scenario where an odd register index is present
in IPSR, the LSB of the register index is utilized as MSB of the bank
number. Observe how we pack reg/bank:
    'IPSR(((reg) << 4U) | (bank), shift, func)' (macro IPnSR)
and how it is read from the device tree source:
    '(RCAR_IPSR(node_id) >> 10U) & 0x1FU' (macro RCAR_PIN_FUNC).

 Finally, this bank is used to obtain the required IPSR offset:
    'PFC_RCAR_IPSR + rcar_func->bank * sizeof(uint32_t)'
in the 'pfc_rcar_set_ipsr' function. For example, if we have the 1 as
a reg index and the 3-rd bank, the resulting offset would be 19,
which is beyond the IPSR range.

Align the IPSR comment with the definition of the 'rcar_pin_func'
structure.

information is obtained from the pin number inside the
'pfc_rcar_get_reg_index' function.

(cherry picked from commit caf9a533301aed7d8213e4f28670cb424de1a9ce)

Original-Note: we can omit the usage of the register index entirely since this
Original-Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>
GitOrigin-RevId: caf9a533301aed7d8213e4f28670cb424de1a9ce
Change-Id: I852887f9e2c731ed2a4cd2dba0799dc380ac5632
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5236632
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>