drivers: display: stm32_ltdc: fix for stm32f429i_disc1
Display is not working on STM32F429i-DISC1 board because
display_blanking_off() needs to be sent to ILI9341 device, but it's sent
to LTDC instead which does not implement it.
This patch adds a LTDC DT property that provides the pHandle of the
display's own controller so that display_blanking_off/on are forwarded to
it when they are called by an application.

(cherry picked from commit 32fd2f57b161529485550199737978e2bb4d2cc4)

Original-Signed-off-by: Abderrahmane Jarmouni <abderrahmane.jarmouni-ext@st.com>
GitOrigin-RevId: 32fd2f57b161529485550199737978e2bb4d2cc4
Change-Id: Iccd77affc9209751d20072f6ac844f4b0de76c62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5280244
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>