boards: arm64: enable ARMV8_A_NS by default for R-Car boards
We need to enable this configuration for all R-Car ARM64 boards.
First and foremost, we definitely should run Zephyr on the boards
in the NS-EL1 state. The EL3 is used for TF-A, EL2 is used for
U-Boot, and Xen in the case when we run Zephyr as Dom-0. The S-EL1
is used for OP-Tee, and S-EL0 is used for OP-Tee apps.

(cherry picked from commit a08604c631138623eb0241cb3d0cdf83bd2595f6)

Original-Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>
GitOrigin-RevId: a08604c631138623eb0241cb3d0cdf83bd2595f6
Change-Id: I961e73f77996481fd7884953c5c65bd619248fa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5279965
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>